imp: Allow already built bundles to exist in /Software while build-and-push-all is working.

+++ software-rebuilder/build-and-push-all.sh
@@ -36,10 +36,16 @@ for software in $(${CAT_BIN} ${_working_state_file} 2>/dev/null); do
             note "Sofin definitions reset for production host type with undefined $(distn "DEVEL")"
     fi
 
-    note "Processing software: $(distn "${software}")"
-    ${SOFIN_BIN} rm ${software}
-    DEVEL=YES ${SOFIN_BIN} deploy ${software} && \
-    ${SED_BIN} -i '' -e "/${software}/d" ${_working_state_file}
-    # ${SOFIN_BIN} rm ${software} && \
+    _indicator="/Software/${software}/$(lowercase "${software}")${DEFAULT_INST_MARK_EXT}"
+    if [ -d "/Software/${software}" -a \
+         -f "${_indicator}" ]; then
+        warn "Found already prebuilt version of software: $(distw "${software}"). Leaving untouched with version: $(distw "$(${CAT_BIN} "${_indicator}" 2>/dev/null)")"
+        ${SED_BIN} -i '' -e "/${software}/d" ${_working_state_file}
+    else
+        note "Processing software: $(distn "${software}")"
+        ${SOFIN_BIN} rm ${software}
+        DEVEL=YES ${SOFIN_BIN} deploy ${software} && \
+        ${SED_BIN} -i '' -e "/${software}/d" ${_working_state_file}
+    fi
     note "--------------------------------"
 done
